requirements.txt
discord.py[voice]
python-dotenv
aiohttp
google-genai
beautifulsoup4
httpx
playwright
psutil
